+++ XolphinApiDotNet/Models/DCVType.cs
@@ -1,5 +1,10 @@
-﻿namespace XolphinApiDotNet.Models
+﻿using Newtonsoft.Json.Converters;
+using System.Text.Json.Serialization;
+
+namespace XolphinApiDotNet.Models
 {
+
+    [JsonConverter(typeof(JsonStringEnumConverter))]
     public enum DCVType
     {
         EMAIL,