+++ XolphinApiDotNet/Requests/Reissue.cs
@@ -29,10 +29,17 @@ namespace XolphinApiDotNet.Requests
         public string Address { get; set; }
         public string Zipcode { get; set; }
         public string City { get; set; }
+        public string Province { get; set; }
+        public string Country { get; set; }        
         public string ApproverFirstName { get; set; }
         public string ApproverLastName { get; set; }
         public string ApproverEmail { get; set; }
         public string ApproverPhone { get; set; }
+        public string ApproverRepresentativeFirstName { get; set; }
+        public string ApproverRepresentativeLastName { get; set; }
+        public string ApproverRepresentativeEmail { get; set; }
+        public string ApproverRepresentativePhone { get; set; }
+        public string ApproverRepresentativePosition { get; set; }        
         public string KVK { get; set; }
         public string Reference { get; set; }
 
@@ -104,6 +111,47 @@ namespace XolphinApiDotNet.Requests
             ApproverPhone = approverPhone;
             return this;
         }
+        public Reissue SetApproverRepresentativeFirstName(string approverRepresentativeFirstName)
+        {
+            ApproverRepresentativeFirstName = approverRepresentativeFirstName;
+            return this;
+        }
+
+        public Reissue SetApproverRepresentativeLastName(string approverRepresentativeLastName)
+        {
+            ApproverRepresentativeLastName = approverRepresentativeLastName;
+            return this;
+        }
+
+        public Reissue SetApproverRepresentativeEmail(string approverRepresentativeEmail)
+        {
+            ApproverRepresentativeEmail = approverRepresentativeEmail;
+            return this;
+        }
+
+        public Reissue SetApproverRepresentativePhone(string approverRepresentativePhone)
+        {
+            ApproverRepresentativePhone = approverRepresentativePhone;
+            return this;
+        }
+
+        public Reissue SetApproverRepresentativePosition(string approverRepresentativePosition)
+        {
+            ApproverRepresentativePosition = approverRepresentativePosition;
+            return this;
+        }
+
+        public Reissue SetProvince(string province)
+        {
+            Province = province;
+            return this;
+        } 
+        
+        public Reissue SetCountry(string country)
+        {
+            Country = country;
+            return this;
+        }        
 
         public Reissue SetKvk(string kvk)
         {