+++ XolphinApiDotNet/Requests/Renew.cs
@@ -32,10 +32,17 @@ namespace XolphinApiDotNet.Requests
         public string Address { get; set; }
         public string Zipcode { get; set; }
         public string City { get; set; }
+        public string Province { get; set; }
+        public string Country { get; set; }        
         public string ApproverFirstName { get; set; }
         public string ApproverLastName { get; set; }
         public string ApproverEmail { get; set; }
         public string ApproverPhone { get; set; }
+        public string ApproverRepresentativeFirstName { get; set; }
+        public string ApproverRepresentativeLastName { get; set; }
+        public string ApproverRepresentativeEmail { get; set; }
+        public string ApproverRepresentativePhone { get; set; }
+        public string ApproverRepresentativePosition { get; set; }        
         public string KVK { get; set; }
         public string Reference { get; set; }
         public string CertificateChainAlternative { get; set; }
@@ -114,6 +121,48 @@ namespace XolphinApiDotNet.Requests
         public Renew SetApproverPhone(string approverPhone)
         {
             ApproverPhone = approverPhone;
+            return this;
+        }
+
+        public Renew SetApproverRepresentativeFirstName(string approverRepresentativeFirstName)
+        {
+            ApproverRepresentativeFirstName = approverRepresentativeFirstName;
+            return this;
+        }
+
+        public Renew SetApproverRepresentativeLastName(string approverRepresentativeLastName)
+        {
+            ApproverRepresentativeLastName = approverRepresentativeLastName;
+            return this;
+        }
+
+        public Renew SetApproverRepresentativeEmail(string approverRepresentativeEmail)
+        {
+            ApproverRepresentativeEmail = approverRepresentativeEmail;
+            return this;
+        }
+
+        public Renew SetApproverRepresentativePhone(string approverRepresentativePhone)
+        {
+            ApproverRepresentativePhone = approverRepresentativePhone;
+            return this;
+        }
+
+        public Renew SetApproverRepresentativePosition(string approverRepresentativePosition)
+        {
+            ApproverRepresentativePosition = approverRepresentativePosition;
+            return this;
+        }
+
+        public Renew SetProvince(string province)
+        {
+            Province = province;
+            return this;
+        } 
+        
+        public Renew SetCountry(string country)
+        {
+            Country = country;
             return this;
         }